refactor: 커뮤니티 상세글 컨트롤러 통합
