+++ src/main/java/com/donguri/jejudorang/domain/community/api/CommunityController.java
@@ -1,12 +1,16 @@
 package com.donguri.jejudorang.domain.community.api;
 
 import com.donguri.jejudorang.domain.community.dto.request.CommunityWriteRequestDto;
+import com.donguri.jejudorang.domain.community.dto.response.ChatDetailResponseDto;
 import com.donguri.jejudorang.domain.community.dto.response.CommunityForModifyResponseDto;
 import com.donguri.jejudorang.domain.community.dto.response.CommunityTypeResponseDto;
+import com.donguri.jejudorang.domain.community.dto.response.PartyDetailResponseDto;
 import com.donguri.jejudorang.domain.community.service.ChatService;
 import com.donguri.jejudorang.domain.community.service.CommunityService;
 import com.donguri.jejudorang.domain.community.service.PartyService;
 import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,8 +25,9 @@ import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.HashMap;
+import java.util.Arrays;
 import java.util.Map;
+import java.util.Optional;
 
 
 @Slf4j
@@ -31,29 +36,22 @@ import java.util.Map;
 public class CommunityController {
 
     private final String kakaoApiKey;
+    private final int viewCookieTime;
 
     @Autowired private final CommunityService communityService;
     @Autowired private final ChatService chatService;
     @Autowired private final PartyService partyService;
 
-    public CommunityController(@Value("${kakao-api-key}") String kakaoApiKey, CommunityService communityService, ChatService chatService, PartyService partyService) {
+    public CommunityController(@Value("${kakao-api-key}") String kakaoApiKey,
+                               @Value("${view.cookie-expire}") int viewCookieTime,
+                               CommunityService communityService, ChatService chatService, PartyService partyService) {
         this.kakaoApiKey = kakaoApiKey;
+        this.viewCookieTime = viewCookieTime;
         this.communityService = communityService;
         this.chatService = chatService;
         this.partyService = partyService;
     }
 
-    // default: api package 내에서만 사용 가능 - getPartyList, getCharList
-    static String convertToProperty(String order) {
-        if (order.equals("recent")) {
-            order = "createdAt";
-        } else if (order.equals("comment")) {
-            order = "comments";
-        } else if (order.equals("bookmark")) {
-            order = "bookmarksCount";
-        }
-        return order;
-    }
 
     /*
     * 글 작성
@@ -125,7 +123,6 @@ public class CommunityController {
     }
 
 
-
     /*
     * 게시글 메인 - 목록 불러오기
     * /community/boards/{type} // parties, chats
@@ -142,6 +139,16 @@ public class CommunityController {
     * String search: 검색어
     * String searchTag: 검색 태그 (A,B,C,...,N개)
     *
+    * > Return Model Attributes
+    * String nowType: 글 목록 타입 - parties, chats
+    * String nowState: 목록 정렬한 모집 상태 - all / recruiting / done - party
+    * String order: 정렬 기준 - recent, comment, bookmark
+    * int nowPage: 지금 페이지
+    * String currentSearchWord: 검색어
+    * String currentSearchTag: 검색 태그
+    *
+    * int allPartyPageCount/allChatPageCount: 목록 전체 페이지 수
+    * Page<Community> partyListDtoPage/chatListDtoPage: 글 데이터
     * */
     @GetMapping("/boards/{type}")
     public String getCommunityList(@PathVariable(name = "type") String type,
@@ -182,6 +189,110 @@ public class CommunityController {
         }
 
         return "/community/communityList";
+    }
+
+    /*
+     * 넘어온 정렬 기준 -> 실제 DB 컬럼명으로 변환
+     * */
+    private String convertToProperty(String order) {
+        switch (order) {
+            case "recent" -> order = "createdAt";
+            case "comment" -> order = "comments";
+            case "bookmark" -> order = "bookmarksCount";
+        }
+        return order;
+    }
+
+
+    /*
+     * 상세글 조회
+     * /community/{type}/{communityId}
+     * GET
+     *
+     * {type} = parties, chats
+     *
+     * */
+    @GetMapping("/boards/{type}/{communityId}")
+    public String getPartyDetail(@PathVariable(name = "type") String type,
+                                 @PathVariable("communityId") Long communityId,
+                                 HttpServletRequest request, HttpServletResponse response,
+                                 Model model) {
+
+        // 쿠키 체크 & 조회수 업데이트 여부 결정 & 조건 충족할 경우 조회수, 쿠키 업데이트
+        checkIsAlreadyReadForUpdateView(communityId, request, response);
+
+        if (type.equals("parties")) {
+            PartyDetailResponseDto foundPartyPost = partyService.getPartyPost(communityId);
+            model.addAttribute("post", foundPartyPost);
+        } else {
+            ChatDetailResponseDto foundChatPost = chatService.getChatPost(communityId);
+            model.addAttribute("post", foundChatPost);
+        }
+
+        model.addAttribute("kakaoApiKey", kakaoApiKey);
+        return "/community/communityDetail";
+    }
+
+    /*
+     * 조회수 중복 필터링 (쿠키 확인)
+     *
+     * # 조회수 증가
+     * 1. 비회원: 쿠키가 아예 없는 경우  [ isRead 쿠키 생성 ]
+     * 2. 회원: 글 조회 목록 쿠키(isRead)가 없는 경우   [ isRead 쿠키 생성 ]
+     * 3. 회원: 글 조회 목록 쿠키(isRead)는 존재하지만, 현재 글 아이디가 포함되지 않은 경우   [ isRead 쿠키 업데이트 ]
+     *
+     * # 조회수, 쿠키 고정
+     * 글 조회 목록 쿠키(isRead)에 현재 글 아이디가 포함될 경우
+     *
+     * */
+    private void checkIsAlreadyReadForUpdateView(Long communityId, HttpServletRequest request, HttpServletResponse response) {
+
+        Optional<Cookie[]> cookies = Optional.ofNullable(request.getCookies());
+
+        // 쿠키가 아예 없거나(비회원) 상세글 조회 목록 쿠키가 없는 경우(회원: 액세스 토큰 쿠키 존재)
+        if (cookies.isEmpty() || Arrays.stream(cookies.get()).filter(cookie -> cookie.getName().equals("isRead")).toList().isEmpty()) {
+
+            Cookie newCookie = new Cookie("isRead", String.valueOf(communityId));
+            updateCookie(response, newCookie);
+            log.info("새로운 쿠키 생성  {} : {}", newCookie.getName(), newCookie.getValue());
+
+            partyService.updatePartyView(communityId);
+            log.info("조회수 증가 완료");
+
+
+        // 상세글 조회 목록 쿠키가 있는 경우
+        } else {
+            Cookie isReadCookie = Arrays.stream(cookies.get())
+                    .filter(coo -> coo.getName().equals("isRead"))
+                    .toList().get(0);
+
+            boolean communityIdExists = Arrays.asList(isReadCookie.getValue().split("/")).contains(String.valueOf(communityId));
+
+            // 현재 communityId가 쿠키 값에 포함되어 있지 않은 경우
+            if (!communityIdExists) {
+                StringBuilder newValueBuilder = new StringBuilder();
+                newValueBuilder.append(isReadCookie.getValue()).append("/").append(communityId);
+
+                isReadCookie.setValue(newValueBuilder.toString());
+                updateCookie(response, isReadCookie);
+                log.info("쿠키에 새로운 communityId 추가: {} -> {}", communityId, newValueBuilder);
+
+                partyService.updatePartyView(communityId);
+                log.info("조회수 증가 완료");
+
+                // 현재 communityId가 쿠키 값에 포함된 경우
+            } else {
+                log.info("이미 조회한 글입니다.");
+            }
+        }
+    }
+
+    // 쿠키 값 설정 후 Response에 추가
+    private void updateCookie(HttpServletResponse response, Cookie newCookie) {
+        newCookie.setHttpOnly(true);
+        newCookie.setMaxAge(viewCookieTime);
+        newCookie.setPath("/");
+        response.addCookie(newCookie);
     }
 
 }
