+++ src/main/resources/templates/community/communityList.html
@@ -144,7 +144,7 @@
             <!--     게시글 카드       -->
             <div class="board-cards-box">
                 <div th:if="${nowType == 'parties'}" th:each="post : ${partyListDtoPage}" class="card w-75 mb-3" style="min-width:100%; max-height:170px;">
-                    <div class="card-body" style="width:100%;" th:onclick="|location.href='@{/community/parties/{communityId}(communityId=${post.id})}'|">
+                    <div class="card-body" style="width:100%;" th:onclick="|location.href='@{/community/boards/parties/{communityId}(communityId=${post.id})}'|">
                         <h5 class="card-title" style="font-weight: bold;" th:text="${post.title}">Card title</h5>
 
                         <div class="card-tags">
@@ -171,7 +171,7 @@
 
 
                 <div th:if="${nowType == 'chats'}" th:each="post : ${chatListDtoPage}" class="card w-75 mb-3" style="min-width:100%; max-height:170px;">
-                    <div class="card-body" style="width:100%;" th:onclick="|location.href='@{/community/chats/{communityId}(communityId=${post.id})}'|">
+                    <div class="card-body" style="width:100%;" th:onclick="|location.href='@{/community/boards/chats/{communityId}(communityId=${post.id})}'|">
                         <h5 class="card-title" style="font-weight: bold;" th:text="${post.title}">Card title</h5>
 
                         <div class="card-tags">
